created a class file to represent the customer object and make handling data returned from the database easier to use.
customer.ts-Routes
changed post route to now call an intermediary method. This method -- createCustomers -- is used to detect the correct code flow to use when the request object contains a single customer and an array of customers.  Based on the data type, the appropriate private method will be called.
added createCustomersFromArray method.  This method generates multiple customer documents from an array of JSON objects that represent a customer object.