userClass.ts: created a class to model the user object.
customers.ts - routes:
modifed the get route to now properly populate the customers object data
modifed the subscription of the get route to now only send the customer data for the user rather than the entire user object.  This change was for security purposes to prevent email / password transmissions.

userRoute.ts
modified the endpoint of the get route to no longer take a searchString parameter.  Will provide a separate route for this later.