customer.ts: modified the create console log to properly describe the action taking place updated javascript doc
transactionRoute.ts:
added javascript docs to the methods
added new patch routes for adding new appointments, adding new payments, deleting appointments, and deleting payments
added new method updatePaymentOrAppointmentData. This method is a powerful and highly flexible method that enables all of the patch routes to be handled in one core set of code.
added enumeration to lock the options the updateType argument to a specific set of options for the updatePaymentOrAppointmentData method.

server.ts:
added the TransactionAPI routes